# NeuroDesk/core/http/endpoints/neurodesk/face_verify.go
package neurodesk

import (
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/mudler/NeuroDesk/core/backend"
	"github.com/mudler/NeuroDesk/core/config"
	"github.com/mudler/NeuroDesk/core/http/middleware"
	"github.com/mudler/NeuroDesk/core/schema"
	"github.com/mudler/NeuroDesk/pkg/model"
	"github.com/mudler/xlog"
)

// FaceVerifyEndpoint compares two images and reports whether they depict the same person.
// @Summary Verify that two images depict the same person.
// @Tags face-recognition
// @Param request body schema.FaceVerifyRequest true "query params"
// @Success 200 {object} schema.FaceVerifyResponse "Response"
// @Router /v1/face/verify [post]
func FaceVerifyEndpoint(cl *config.ModelConfigLoader, ml *model.ModelLoader, appConfig *config.ApplicationConfig) echo.HandlerFunc {
	return func(c echo.Context) error {
		input, ok := c.Get(middleware.CONTEXT_LOCALS_KEY_NEURODESK_REQUEST).(*schema.FaceVerifyRequest)
		if !ok || input.Model == "" {
			return echo.ErrBadRequest
		}
		cfg, ok := c.Get(middleware.CONTEXT_LOCALS_KEY_MODEL_CONFIG).(*config.ModelConfig)
		if !ok || cfg == nil {
			return echo.ErrBadRequest
		}

		img1, err := decodeImageInput(input.Img1)
		if err != nil {
			return err
		}
		img2, err := decodeImageInput(input.Img2)
		if err != nil {
			return err
		}

		xlog.Debug("FaceVerify", "model", cfg.Name, "backend", cfg.Backend)
		res, err := backend.FaceVerify(c.Request().Context(), img1, img2, input.Threshold, input.AntiSpoofing, ml, appConfig, *cfg)
		if err != nil {
			return mapBackendError(err)
		}

		out := schema.FaceVerifyResponse{
			Verified:   res.GetVerified(),
			Distance:   res.GetDistance(),
			Threshold:  res.GetThreshold(),
			Confidence: res.GetConfidence(),
			Model:      res.GetModel(),
			Img1Area: schema.FacialArea{
				X: res.GetImg1Area().GetX(),
				Y: res.GetImg1Area().GetY(),
				W: res.GetImg1Area().GetW(),
				H: res.GetImg1Area().GetH(),
			},
			Img2Area: schema.FacialArea{
				X: res.GetImg2Area().GetX(),
				Y: res.GetImg2Area().GetY(),
				W: res.GetImg2Area().GetW(),
				H: res.GetImg2Area().GetH(),
			},
			ProcessingTimeMs: res.GetProcessingTimeMs(),
		}
		if input.AntiSpoofing {
			img1IsReal := res.GetImg1IsReal()
			img1Score := res.GetImg1AntispoofScore()
			img2IsReal := res.GetImg2IsReal()
			img2Score := res.GetImg2AntispoofScore()
			out.Img1IsReal = &img1IsReal
			out.Img1AntispoofScore = &img1Score
			out.Img2IsReal = &img2IsReal
			out.Img2AntispoofScore = &img2Score
		}
		return c.JSON(http.StatusOK, out)
	}
}