# NeuroDesk/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The NeuroDesk Rest API.",
        "title": "NeuroDesk API",
        "contact": {
            "name": "NeuroDesk",
            "url": "https://localai.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://raw.githubusercontent.com/mudler/NeuroDesk/master/LICENSE"
        },
        "version": "2.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/3d/remesh": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "model/gltf-binary"
                ],
                "tags": [
                    "3d"
                ],
                "summary": "Applies watertight print remeshing to an existing 3D asset.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "3D model name",
                        "name": "model",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Source GLB",
                        "name": "mesh",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Detail size as percent of the source bounding-box diagonal (0.35–2.5; default 0.5)",
                        "name": "detail",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Remeshed GLB",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
        "/v1/face/verify": {
            "post": {
                "tags": [
                    "face-recognition"
                ],
                "summary": "Verify that two images depict the same person.",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.FaceVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response",
                        "schema": {
                            "$ref": "#/definitions/schema.FaceVerifyResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/voice/identify": {
            "post": {
                "tags": [
                    "voice-recognition"
                ],
                "summary": "Identify a speaker against the registered database (1:N recognition).",
                "parameters": [
                    {
                        "description": "query params",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.VoiceIdentifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response",
                        "schema": {
                            "$ref": "#/definitions/schema.VoiceIdentifyResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.FaceVerifyRequest": {
            "type": "object",
            "properties": {
                "anti_spoofing": {
                    "type": "boolean"
                },
                "img1": {
                    "type": "string"
                },
                "img2": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "threshold": {
                    "type": "number"
                }
            }
        },
        "schema.FaceVerifyResponse": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "distance": {
                    "type": "number"
                },
                "img1_antispoof_score": {
                    "type": "number"
                },
                "img1_area": {
                    "$ref": "#/definitions/schema.FacialArea"
                },
                "img1_is_real": {
                    "description": "Liveness fields are only populated when the request set\nanti_spoofing=true. Pointers keep them fully absent from the\nJSON response otherwise, so callers can tell \"not checked\"\napart from \"checked and fake\" (which would collapse to zero\nvalues with plain bool+omitempty).",
                    "type": "boolean"
                },
                "img2_antispoof_score": {
                    "type": "number"
                },
                "img2_area": {
                    "$ref": "#/definitions/schema.FacialArea"
                },
                "img2_is_real": {
                    "type": "boolean"
                },
                "model": {
                    "type": "string"
                },
                "processing_time_ms": {
                    "type": "number"
                },
                "threshold": {
                    "type": "number"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
        "schema.FacialArea": {
            "type": "object",
            "properties": {
                "h": {
                    "type": "number"
                },
                "w": {
                    "type": "number"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
...
        "schema.VoiceIdentifyMatch": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "distance": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "match": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "schema.VoiceIdentifyRequest": {
            "type": "object",
            "properties": {
                "audio": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "store": {
                    "type": "string"
                },
                "threshold": {
                    "type": "number"
                },
                "top_k": {
                    "type": "integer"
                }
            }
        },
        "schema.VoiceIdentifyResponse": {
            "type": "object",
            "properties": {
                "matches": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.VoiceIdentifyMatch"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}